requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
groq